+++ Functions.Templates/Templates/HttpTrigger-CSharp-Isolated-NetFx/.template.config/vs-2017.3.host.json
@@ -1,0 +1,29 @@
+{
+  "name": {
+    "text": "Http trigger",
+    "package": "{E9BCFE2B-304A-4D7F-836F-F3A4FB1A1451}",
+     "id": "2050"
+  },
+  "description": {
+    "text": "A C# function that will be run whenever it receives an HTTP request",
+    "package": "{E9BCFE2B-304A-4D7F-836F-F3A4FB1A1451}",
+     "id": "2051"
+  },
+  "order": 100,
+  "icon": "vs-2017.3/httpTrigger.png",
+  "learnMoreLink": "https://go.microsoft.com/fwlink/?linkid=848145",
+  "uiFilters": [
+    "functions"
+  ],
+  "symbolInfo": [
+    {
+      "id": "AccessRights",
+      "name": {
+        "id": "2052",
+        "text": "Authorization level",
+        "package": "{E9BCFE2B-304A-4D7F-836F-F3A4FB1A1451}"
+      },
+      "isVisible": true
+    }
+  ]
+}
